Refactored artist blueprint
